Sensorslib: Move `mockSensorEvent` in `Mocks` file
This commit will move the method `mockSensorEvent` inside the `Mocks` file
since it is used to create some mocks.
